+++ spec/asm/shellcode_spec.rb
@@ -1,3 +1,5 @@
+# encoding: US-ASCII
+
 require 'spec_helper'
 require 'ronin/asm/shellcode'
 
@@ -9,17 +11,17 @@ describe ASM::Shellcode do
         push  eax
         push  0x68732f2f
         push  0x6e69622f
-        mov   esp,  ebx
+        mov   ebx, esp
         push  eax
         push  ebx
-        mov   esp,  ecx
+        mov   ecx,  esp
         xor   edx,  edx
-        mov   0xb,  al
+        mov   al,   0xb
         int   0x80
       end
     end
 
-    let(:shellcode) { "f1\xC0fPfh//shfh/binf\x89\xE3fPfSf\x89\xE1f1\xD2\xB0\v\xCD\x80" }
+    let(:shellcode) { "1\xC0Ph//shh/bin\x89\xE3PS\x89\xE11\xD2\xB0\v\xCD\x80" }
 
     it "assemble down to raw machine code" do
       subject.assemble.should == shellcode
@@ -28,6 +30,12 @@ describe ASM::Shellcode do
     context "with :syntax is :intel" do
       it "assemble down to raw machine code" do
         subject.assemble(syntax: :intel).should == shellcode
+      end
+    end
+
+    context "with :syntax is :att" do
+      it "assemble down to raw machine code" do
+        subject.assemble(syntax: :att).should == shellcode
       end
     end
   end